+++ crates/ars-dom/src/lib.rs
@@ -24,7 +24,6 @@ pub mod portal;
 pub mod positioning;
 mod scroll;
 pub mod scroll_lock;
-pub mod url;
 pub mod z_index;
 
 #[cfg(feature = "web")]
@@ -65,7 +64,6 @@ pub use scroll_lock::{
 };
 #[cfg(all(feature = "web", target_arch = "wasm32"))]
 pub use scroll_lock::{needs_ios_workaround, scrollbar_width};
-pub use url::{SafeUrl, UnsafeUrlError, is_safe_url, sanitize_url};
 pub use z_index::{ZIndexAllocator, next_z_index, reset_z_index, supports_top_layer};
 
 /// Describes the platform capabilities available to the current runtime.
