+++ crates/ars-forms/src/validation/error.rs
@@ -5,7 +5,7 @@
 
 use ars_i18n::Locale;
 
-use crate::form_messages::FormMessages;
+use crate::form::Messages;
 
 /// A single validation failure.
 #[derive(Clone, Debug, PartialEq)]
@@ -19,7 +19,7 @@ pub struct Error {
 
 impl Error {
     /// Creates a required-field validation error with a localized message.
-    pub fn required(messages: &FormMessages, locale: &Locale) -> Self {
+    pub fn required(messages: &Messages, locale: &Locale) -> Self {
         Self {
             message: (messages.required_error)(locale),
             code: ErrorCode::Required,
@@ -27,7 +27,7 @@ impl Error {
     }
 
     /// Creates a minimum-length validation error with a localized message.
-    pub fn min_length(min: usize, messages: &FormMessages, locale: &Locale) -> Self {
+    pub fn min_length(min: usize, messages: &Messages, locale: &Locale) -> Self {
         Self {
             message: (messages.min_length_error)(min, locale),
             code: ErrorCode::MinLength(min),
@@ -35,7 +35,7 @@ impl Error {
     }
 
     /// Creates a maximum-length validation error with a localized message.
-    pub fn max_length(max: usize, messages: &FormMessages, locale: &Locale) -> Self {
+    pub fn max_length(max: usize, messages: &Messages, locale: &Locale) -> Self {
         Self {
             message: (messages.max_length_error)(max, locale),
             code: ErrorCode::MaxLength(max),
@@ -43,7 +43,7 @@ impl Error {
     }
 
     /// Creates a pattern validation error with a localized message.
-    pub fn pattern(pattern: impl Into<String>, messages: &FormMessages, locale: &Locale) -> Self {
+    pub fn pattern(pattern: impl Into<String>, messages: &Messages, locale: &Locale) -> Self {
         Self {
             message: (messages.pattern_error)(locale),
             code: ErrorCode::Pattern(pattern.into()),
@@ -51,7 +51,7 @@ impl Error {
     }
 
     /// Creates a minimum-value validation error with a localized message.
-    pub fn min(min: f64, messages: &FormMessages, locale: &Locale) -> Self {
+    pub fn min(min: f64, messages: &Messages, locale: &Locale) -> Self {
         Self {
             message: (messages.min_error)(min, locale),
             code: ErrorCode::Min(min),
@@ -59,7 +59,7 @@ impl Error {
     }
 
     /// Creates a maximum-value validation error with a localized message.
-    pub fn max(max: f64, messages: &FormMessages, locale: &Locale) -> Self {
+    pub fn max(max: f64, messages: &Messages, locale: &Locale) -> Self {
         Self {
             message: (messages.max_error)(max, locale),
             code: ErrorCode::Max(max),
@@ -67,7 +67,7 @@ impl Error {
     }
 
     /// Creates an email validation error with a localized message.
-    pub fn email(messages: &FormMessages, locale: &Locale) -> Self {
+    pub fn email(messages: &Messages, locale: &Locale) -> Self {
         Self {
             message: (messages.email_error)(locale),
             code: ErrorCode::Email,
@@ -75,7 +75,7 @@ impl Error {
     }
 
     /// Creates a step validation error with a localized message.
-    pub fn step(step: f64, messages: &FormMessages, locale: &Locale) -> Self {
+    pub fn step(step: f64, messages: &Messages, locale: &Locale) -> Self {
         Self {
             message: (messages.step_error)(step, locale),
             code: ErrorCode::Step(step),
@@ -83,7 +83,7 @@ impl Error {
     }
 
     /// Creates a URL validation error with a localized message.
-    pub fn url(messages: &FormMessages, locale: &Locale) -> Self {
+    pub fn url(messages: &Messages, locale: &Locale) -> Self {
         Self {
             message: (messages.url_error)(locale),
             code: ErrorCode::Url,
@@ -207,7 +207,7 @@ mod tests {
     use ars_i18n::locales;
 
     use super::*;
-    use crate::form_messages::FormMessages;
+    use crate::form::Messages;
 
     #[test]
     fn valid_result_is_server() {
@@ -298,7 +298,7 @@ mod tests {
 
     #[test]
     fn error_required_factory() {
-        let err = Error::required(&FormMessages::default(), &locales::en());
+        let err = Error::required(&Messages::default(), &locales::en());
 
         assert_eq!(err.message, "This field is required");
         assert_eq!(err.code, ErrorCode::Required);
@@ -306,7 +306,7 @@ mod tests {
 
     #[test]
     fn error_min_length_factory() {
-        let err = Error::min_length(3, &FormMessages::default(), &locales::en());
+        let err = Error::min_length(3, &Messages::default(), &locales::en());
 
         assert_eq!(err.message, "Must be at least 3 characters");
         assert_eq!(err.code, ErrorCode::MinLength(3));
@@ -314,7 +314,7 @@ mod tests {
 
     #[test]
     fn error_max_length_factory() {
-        let err = Error::max_length(8, &FormMessages::default(), &locales::en());
+        let err = Error::max_length(8, &Messages::default(), &locales::en());
 
         assert_eq!(err.message, "Must be at most 8 characters");
         assert_eq!(err.code, ErrorCode::MaxLength(8));
@@ -322,7 +322,7 @@ mod tests {
 
     #[test]
     fn error_pattern_factory() {
-        let err = Error::pattern(r"^[a-z]+$", &FormMessages::default(), &locales::en());
+        let err = Error::pattern(r"^[a-z]+$", &Messages::default(), &locales::en());
 
         assert_eq!(err.message, "Invalid format");
         assert_eq!(err.code, ErrorCode::Pattern(String::from(r"^[a-z]+$")));
@@ -333,7 +333,7 @@ mod tests {
         let pattern = String::from(r"^[a-z]+$");
         let errors = Errors(vec![Error::pattern(
             pattern.clone(),
-            &FormMessages::default(),
+            &Messages::default(),
             &locales::en(),
         )]);
 
@@ -342,7 +342,7 @@ mod tests {
 
     #[test]
     fn error_min_factory() {
-        let err = Error::min(2.5, &FormMessages::default(), &locales::en());
+        let err = Error::min(2.5, &Messages::default(), &locales::en());
 
         assert_eq!(err.message, "Must be at least 2.5");
         assert_eq!(err.code, ErrorCode::Min(2.5));
@@ -350,7 +350,7 @@ mod tests {
 
     #[test]
     fn error_max_factory() {
-        let err = Error::max(9.5, &FormMessages::default(), &locales::en());
+        let err = Error::max(9.5, &Messages::default(), &locales::en());
 
         assert_eq!(err.message, "Must be at most 9.5");
         assert_eq!(err.code, ErrorCode::Max(9.5));
@@ -358,7 +358,7 @@ mod tests {
 
     #[test]
     fn error_email_factory() {
-        let err = Error::email(&FormMessages::default(), &locales::en());
+        let err = Error::email(&Messages::default(), &locales::en());
 
         assert_eq!(err.message, "Must be a valid email address");
         assert_eq!(err.code, ErrorCode::Email);
@@ -366,7 +366,7 @@ mod tests {
 
     #[test]
     fn error_step_factory() {
-        let err = Error::step(0.25, &FormMessages::default(), &locales::en());
+        let err = Error::step(0.25, &Messages::default(), &locales::en());
 
         assert_eq!(
             err.message,
@@ -377,7 +377,7 @@ mod tests {
 
     #[test]
     fn error_url_factory() {
-        let err = Error::url(&FormMessages::default(), &locales::en());
+        let err = Error::url(&Messages::default(), &locales::en());
 
         assert_eq!(err.message, "Please enter a valid URL.");
         assert_eq!(err.code, ErrorCode::Url);
