+++ spec/dioxus-components/utility/field.md
@@ -237,7 +237,8 @@ pub struct FieldSketchProps {
 
 #[component]
 pub fn Field(props: FieldSketchProps) -> Element {
-    let machine = use_machine::<field::FieldComponentMachine>(field::Props::default());
+    let machine =
+        use_machine::<field::component::Machine>(field::component::Props::default());
     let root_attrs = machine.derive(|api| api.root_attrs());
     use_context_provider(|| Context(machine));
 
@@ -265,7 +266,7 @@ pub fn Label(props: LabelSketchProps) -> Element {
 ## 25. Reference Implementation Skeleton
 
 ```rust
-let machine = use_machine::<field::Machine>(props);
+let machine = use_machine::<field::component::Machine>(props);
 let ids = derive_stable_field_ids(machine);
 let machine_ctx = publish_field_machine_context(machine);
 let inherited_ctx = publish_field_context(machine, parent_ctx);
