+++ spec/dioxus-components/utility/fieldset.md
@@ -226,7 +226,9 @@ pub struct FieldsetSketchProps {
 
 #[component]
 pub fn Fieldset(props: FieldsetSketchProps) -> Element {
-    let machine = use_machine::<fieldset::Machine>(fieldset::Props::default());
+    let machine = use_machine::<fieldset::component::Machine>(
+        fieldset::component::Props::default(),
+    );
     let root_attrs = machine.derive(|api| api.root_attrs());
 
     use_context_provider(|| Context(machine));
@@ -248,7 +250,7 @@ pub fn Fieldset(props: FieldsetSketchProps) -> Element {
 ## 25. Reference Implementation Skeleton
 
 ```rust
-let machine = use_machine::<fieldset::Machine>(props);
+let machine = use_machine::<fieldset::component::Machine>(props);
 let ids = derive_stable_fieldset_ids(machine);
 publish_fieldset_machine_context(machine);
 publish_inherited_field_ctx(machine);
