+++ spec/leptos-components/utility/field.md
@@ -193,7 +193,8 @@ use leptos::prelude::*;
 
 #[component]
 pub fn Field(children: Children) -> impl IntoView {
-    let machine = use_machine::<field::FieldComponentMachine>(field::Props::default());
+    let machine =
+        use_machine::<field::component::Machine>(field::component::Props::default());
     let root_attrs = machine.derive(|api| api.root_attrs());
     provide_context(Context(machine));
 
@@ -224,7 +225,7 @@ pub fn Description(children: Children) -> impl IntoView {
 ## 25. Reference Implementation Skeleton
 
 ```rust
-let machine = use_machine::<field::Machine>(props);
+let machine = use_machine::<field::component::Machine>(props);
 let ids = derive_stable_field_ids(machine);
 let machine_ctx = publish_field_machine_context(machine);
 let inherited_ctx = publish_field_context(machine, parent_ctx);
